requirements.txt
vk-api==11.9.9
python-dotenv==1.0.0
aiohttp==3.9.3
aiolimiter
tenacity
aiocache